Combine duplicate game teams manager tests
