+++ test/game_teams_manager_test.rb
@@ -87,4 +87,21 @@ class GameTeamsManagerTest < Minitest::Test
     assert_equal "John Tortorella", game_team_manager.worst_coach(game_ids)
   end
 
+  def test_team_id_highest_average_goals_all_full_file
+    skip
+    path = "./data/game_teams.csv"
+    game_teams_manager = GameTeamsManager.new(path)
+
+    assert_equal "54", game_teams_manager.team_id_highest_average_goals_all
+  end
+
+  def test_total_goals_by_team_dummy_file
+    path = "./fixture/game_teams_dummy15.csv"
+    game_teams_manager = GameTeamsManager.new(path)
+
+    expected = {"3"=>8, "6"=>21, "5"=>2}
+
+    assert_equal expected, game_teams_manager.total_goals_by_team
+  end
+
 end
